We need to be able to run checks on files changed this commit.
There is also an option to choose to compare against a
custom commitish, vs the last.

This introduces the allow config, which doubles
as a thing to set for a scm diff check.

Taking away module vile, to make tests work, will
find way to use let again.

Also getting tests up and running for git module and
placeholders for others.