It is more delightful to have code snippets by default.
Note: this possibly creates a slightly negative UX
with users who may not want snippets of their code
uploaded, and who are unpleasantly surprised as a result.
However, services akin to vile.io clone repos and a
precedent exists, so IMO this should not be that surprising.

Even so, you can disable it with a flag, and
can easily delete any commit via the web
interface, including all of its data and snippets.