We need to make Vile more useful as a pure lib.
This includes:

* Use loglevel and chalk vs minilog for logging

* Expose new logger and let lib users disable logging

* Support propagating color to vile logger in sub processes

* Log to stderr wherever possible

* Start using error control flow so plugin/* modules don't
touch process.exit themselves.

* `vile analyze` is more intuitive of a command than `vile punish`.

This was something I made when I prototyped the project and has
been with it ever since. As much as the theme of punishing is
still apparent, this feels more professional, IMO.

* The plugin exec method should not have duplicate plugin opts.

It was a left over legacy API, and needs to be upgraded.

Since no one uses the library itself, this can be done without
any issues.